% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_destinations}
\alias{create_destinations}
\title{Create Output Points}
\usage{
create_destinations(x)
}
\arguments{
\item{x}{a string containing a relative file path}
}
\description{
Take a string containing a file path and validate the existence of all paths
and create directories where applicable.
}
