% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_blank_lines}
\alias{format_blank_lines}
\title{Format sequential blank lines}
\usage{
format_blank_lines(x)
}
\arguments{
\item{x}{a character}
}
\description{
If there are two blank lines in a row, remove the second item
}
