% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_leading_whitespace}
\alias{format_leading_whitespace}
\title{Format Leading Whitespace}
\usage{
format_leading_whitespace(x)
}
\arguments{
\item{x}{a string from an md file}
}
\description{
In many text blocks, there a leading white space is added. This function
removes it!
}
\examples{
format_leading_whitespace("    This is a cool string!   ")

}
