% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_md_table}
\alias{format_md_table}
\title{Format Markdown Table}
\usage{
format_md_table(x)
}
\arguments{
\item{x}{a string or array containing markdown}
}
\description{
Fix markdown table headers, table alignments, and argument cells.
Remove extra whitespacing and multiple backticks. Enclose rows in
with a vertical line.
}
