% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_path}
\alias{format_path}
\title{Format Path}
\usage{
format_path(x)
}
\arguments{
\item{x}{a string containing a file path}
}
\description{
Remove trailing forward slash in file paths.
}
\examples{

format_path("path/to/doc/")

}
