% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_rd_path}
\alias{format_rd_path}
\title{Format Rd File name}
\usage{
format_rd_path(x)
}
\arguments{
\item{x}{a file name}
}
\description{
Remove the file extension from Rd files.
}
\examples{

format_rd_path("man/myfile.Rd")

}
