% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_yaml_keywords}
\alias{format_yaml_keywords}
\title{Format YAML keywords}
\usage{
format_yaml_keywords(x)
}
\arguments{
\item{x}{an Rd keyword array}
}
\description{
Format Rd keywords (post-conversion), into a JS formatted string.
}
\examples{
format_yaml_keywords(c("I", "think", "R", "is", "cool"))

}
