% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_yaml_text}
\alias{format_yaml_text}
\title{Format YAML Text}
\usage{
format_yaml_text(x)
}
\arguments{
\item{x}{an Rd string}
}
\description{
Converted Rd files may not always render YAML markdown. This function
rendered Rd into Markdown and ensures YAML is quoted.
}
\examples{
format_yaml_text("This is a \\\\bold{cool} string")

}
