% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Rd to Md Workflow}
\value{
Convert Rd files to Markdown files
}
\description{
As R6 class for converting Rd files to markdown with YAML headers. This may
be useful if you wish to use package documentation in static site generators
outside of the R ecosystem (e.g., React, Vue, Svelte, Gatsby, etc.). By
default, Rd files are rendered into their own directory with an independent
\code{index.md} file. The Rd name is parsed and set as the child directory name.

For example, if you have the file \code{man/my_cool_function.Rd}, the output
path and file will be: \code{my_cool_function/index.md}

The entry point will always be \code{man}.
}
\keyword{convert}
\keyword{markdown}
\keyword{rd}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{entry}}{path to Rd files}

\item{\code{destDir}}{output directory}

\item{\code{files}}{list of entry files and output paths}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{convert$new()}}
\item \href{#method-set_entries}{\code{convert$set_entries()}}
\item \href{#method-set_destinations}{\code{convert$set_destinations()}}
\item \href{#method-convert_rds}{\code{convert$convert_rds()}}
\item \href{#method-add_yaml}{\code{convert$add_yaml()}}
\item \href{#method-clone}{\code{convert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new converter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$new(destDir = "src/pages")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{destDir}}{output directory for markdown files}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_entries"></a>}}
\if{latex}{\out{\hypertarget{method-set_entries}{}}}
\subsection{Method \code{set_entries()}}{
Set entry points and output file paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$set_entries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_destinations"></a>}}
\if{latex}{\out{\hypertarget{method-set_destinations}{}}}
\subsection{Method \code{set_destinations()}}{
set destinitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$set_destinations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_rds"></a>}}
\if{latex}{\out{\hypertarget{method-convert_rds}{}}}
\subsection{Method \code{convert_rds()}}{
batch convert Rd files to markdown files with YAML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$convert_rds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_yaml"></a>}}
\if{latex}{\out{\hypertarget{method-add_yaml}{}}}
\subsection{Method \code{add_yaml()}}{
inject YAML from Rd files into rendered markdown files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$add_yaml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
