% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Rd to Md Workflow}
\value{
Convert Rd files to Markdown files
}
\description{
`convert` is an R6 class for converting Rd files to markdown for use in
static site generators. This may be useful if you wish to use non-R web
frameworks for generating package documentation or if you want to add
you package man files to an existing web project in markdown format. This
workflow is optimized for the Gatsby JS theme `rocket-docs`, but it can be
used for as a generic renderer.

To get started, create a new instance of the converter. It is also
recommended to convert Rd files into an existing project.

```r
pkg <- rdConvert::convert$new()
```

Next, configure the convert to fit your project. The argument `site_dir` is
used to define the base path for the static site. It is recommended to
create a nested project within your R project. `dest_dir` is the location
that the convert should use to save the markdown files. If you are using
RocketDocs, this is `src/docs/usage/` (no need to supply the site_dir here).

```r
pkg$config(
  project_name = "rdConvert",
  site_dir = "gatsby",
  dest_dir = "src/docs/usage/",
  repo_name = "rdConvert",
  repo_url = "https://github.com/davidruvolo51/rdConvert"
)
```

If `site_dir` is not present you, `config` will through a warning. If you
are using RocketDocs, you can set the path of the `sidebar.yml` file via the
`sidebar_yml_path` argument. By default the location is
`src/config/sidebar.yml`. You can adjust the configuration at any point.

Next, create the entry and output points. This will collate all `man` files
and generate the output path using the `site_dir` and `dest_dir` arguments
defined in by the `$config` function. If you are using RocketDocs, the
sidebar data `labels` and `links` are generated automatically.

```r
pkg$set_entries()   # build
pkg$entries         # view
```

Once the entry and output points are created, run `$set_destinations`. This
function will verify that all directories in a path exist. If a location
does not exist, the folder will be created.

```r
pkg$set_destinations()
```

The returned md file may not always be perfect. There may be extra spaces
and multiple blank lines. Table layouts may not render appropriately. The
function `$format_markdown()` provides some surface-level formatting for
post-converted markdown files.

```r
pkg$format_markdown()
```

If you want to generate YAML for the files, use the `$add_yaml()` function.
This function isn't perfect. If you are using RocketDocs it will cause the
title to render twice as it was rendered via `$convert_rds`. I'm still
working on this feature.

```r
pkg$add_yaml()
```

If you are using RocketDocs, I've included two methods for generating the
`sidebar.yml`. `set_sidebar_yml` will generate the markup using the
`$entries` object. The yml configuration file can be saved using the
`$save_sidebar_yml` function.

```r
pkg$save_sidebar_yml()
```

### Tips for Success

This workflow isn't perfect, but it will help provide a basic workflow for
exporting Rd files for use in non-R web projects. Here are some tips for
success.

- **Create static site project first**: It is a good idea to create the
    static site before you convert the Rd files. This will ensure the
    output paths are initiated before you build the paths.
- **Turn of Markdown support**: While experimenting with this workflow,
    I noticed that roxygen2 was rendering markdown. To display this,
    use `@noMd` in your R files or disable it package wide in the
    `DESCRIPTION >> Roxygen: list(markdown = TRUE)`.

During active development, rebuilding Rd files and rerunning the steps
outlined above, can be a bit tidious. To make this process a bit easier,
run the method `$rebuild`. This will run `devtools::check_man` and
`$convert_rds` and `$format_markdown`. Set `set_entries` to `TRUE` if
you have the Rd files were added or removed. You can also auto add YAMLs
by setting `add_yaml` to TRUE.

```r
pkg$rebuild(set_entries = TRUE, add_yaml = FALSE)
```
}
\author{
dcruvolo
}
\keyword{convert}
\keyword{markdown}
\keyword{rd}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{project_name}}{project name}

\item{\code{site_dir}}{location of the of the static site directory}

\item{\code{entry_dir}}{path to Rd files}

\item{\code{dest_dir}}{output directory}

\item{\code{entries}}{list of entry files and output paths}

\item{\code{file_format}}{save file as md or mdx}

\item{\code{yml}}{configuration file for sidebar.yml}

\item{\code{repo_name}}{name of the repo (e.g., Github, GitLab, etc.)}

\item{\code{repo_url}}{address repo (e.g., GitHub, GitLab, etc.)}

\item{\code{sidebar_yml_path}}{path to RocketDocs sidebar.yml}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-config}{\code{convert$config()}}
\item \href{#method-set_entries}{\code{convert$set_entries()}}
\item \href{#method-set_destinations}{\code{convert$set_destinations()}}
\item \href{#method-convert_rds}{\code{convert$convert_rds()}}
\item \href{#method-format_markdown}{\code{convert$format_markdown()}}
\item \href{#method-add_yaml}{\code{convert$add_yaml()}}
\item \href{#method-set_sidebar_yml}{\code{convert$set_sidebar_yml()}}
\item \href{#method-save_sidebar_yml}{\code{convert$save_sidebar_yml()}}
\item \href{#method-rebuild}{\code{convert$rebuild()}}
\item \href{#method-clone}{\code{convert$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-config"></a>}}
\if{latex}{\out{\hypertarget{method-config}{}}}
\subsection{Method \code{config()}}{
Configure a new converter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$config(
  project_name,
  site_dir = "gatsby",
  dest_dir = "src/docs/usage",
  file_format = "md",
  repo_name = NULL,
  repo_url = NULL,
  sidebar_yml_path = "src/config/sidebar.yml"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_name}}{name of the project}

\item{\code{site_dir}}{location of the static site directory}

\item{\code{dest_dir}}{output directory for markdown files (excl. site dir)}

\item{\code{file_format}}{save file as md or mdx}

\item{\code{repo_name}}{name of the repo (e.g., Github, GitLab, etc.)}

\item{\code{repo_url}}{address repo (e.g., GitHub, GitLab, etc.)}

\item{\code{sidebar_yml_path}}{path to sidebar yml config (excl. site dir)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_entries"></a>}}
\if{latex}{\out{\hypertarget{method-set_entries}{}}}
\subsection{Method \code{set_entries()}}{
Set entry points and output file paths
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$set_entries()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_destinations"></a>}}
\if{latex}{\out{\hypertarget{method-set_destinations}{}}}
\subsection{Method \code{set_destinations()}}{
set destinitions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$set_destinations()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-convert_rds"></a>}}
\if{latex}{\out{\hypertarget{method-convert_rds}{}}}
\subsection{Method \code{convert_rds()}}{
batch convert Rd files to markdown files with YAML
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$convert_rds()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-format_markdown"></a>}}
\if{latex}{\out{\hypertarget{method-format_markdown}{}}}
\subsection{Method \code{format_markdown()}}{
format markdown files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$format_markdown()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_yaml"></a>}}
\if{latex}{\out{\hypertarget{method-add_yaml}{}}}
\subsection{Method \code{add_yaml()}}{
inject YAML from Rd files into rendered markdown files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$add_yaml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_sidebar_yml"></a>}}
\if{latex}{\out{\hypertarget{method-set_sidebar_yml}{}}}
\subsection{Method \code{set_sidebar_yml()}}{
generate sidebar yml
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$set_sidebar_yml()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{home}}{display home link in sidebar}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_sidebar_yml"></a>}}
\if{latex}{\out{\hypertarget{method-save_sidebar_yml}{}}}
\subsection{Method \code{save_sidebar_yml()}}{
write yml config
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$save_sidebar_yml()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-rebuild"></a>}}
\if{latex}{\out{\hypertarget{method-rebuild}{}}}
\subsection{Method \code{rebuild()}}{
rebuild R man files and convert to markdown
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$rebuild(entries = FALSE, yaml = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entries}}{If TRUE, entry and points will be rebuilt}

\item{\code{yaml}}{If TRUE, yamls will be updated}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{convert$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
