% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_blank_lines}
\alias{format_blank_lines}
\title{Format Sequential blank lines}
\usage{
format_blank_lines(x)
}
\arguments{
\item{x}{a character}
}
\description{
In some instances, there may be more than one blank line in between
paragraphs. This function ensures there is only 1 blank line.

```r
format_blank_lines(c("This has", "", "", "", "multiple", "", "","", "", "blank lines"))
```
}
