% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_md_table}
\alias{format_md_table}
\title{Format Markdown Table}
\usage{
format_md_table(x)
}
\arguments{
\item{x}{a string or array containing markdown}
}
\description{
In converted markdown files, the markup can be a bit funny and will render
properly. This is caused by extra spaces, no spaces between column
characters, and multiple backticks around arguments. This function cleans
markdown tables.
}
