% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{format_yaml_keywords}
\alias{format_yaml_keywords}
\title{Format keywords}
\usage{
format_yaml_keywords(x)
}
\arguments{
\item{x}{an Rd keyword array}
}
\description{
Collapse Rd keywords array into JS array
}
